2 funman service files = 1
